+++ plugins/gs/src/components/HelmReleaseDetails/HelmReleaseDetails.tsx
@@ -13,6 +13,7 @@ import {
 } from "../../model/services/mapi/helmv2beta1";
 import { Heading } from "../UI/Heading";
 import { formatVersion } from "../utils/helpers";
+import { GitOpsUILink } from "../UI/GitOpsUILink/GitOpsUILink";
 
 type HelmReleaseDetailsProps = {
   installationName: string;
@@ -61,21 +62,34 @@ export const HelmReleaseDetails = ({
   
   return (
     <div>
-      <Box marginBottom={4}>
-        <StructuredMetadataList metadata={{
-          'Installation': installationName,
-          'Cluster': clusterName === '' ? 'n/a' : clusterName,
-        }} />
+      <Box display='flex' marginBottom={2}>
+        <GitOpsUILink
+          installationName={installationName}
+          clusterName={installationName}
+          kind='helmrelease'
+          name={name}
+          namespace={namespace}
+          text='Open the GitOps UI for this resource'
+        />
       </Box>
       
       <Grid container direction="column">
         <Grid item>
-          <RevisionDetails
-            lastAppliedRevision={lastAppliedRevision}
-            lastAttemptedRevision={lastAttemptedRevision}
-            sourceLocation={sourceLocation}
-          />
+          <Card>
+            <CardContent>
+              <StructuredMetadataList metadata={{
+                'Installation': installationName,
+                'Cluster': clusterName ? clusterName : 'n/a',
+              }} />
+            </CardContent>
+          </Card>
         </Grid>
+
+        <RevisionDetails
+          lastAppliedRevision={lastAppliedRevision}
+          lastAttemptedRevision={lastAttemptedRevision}
+          sourceLocation={sourceLocation}
+        />
 
         <Grid item>
           <HelmReleaseDetailsStatus helmrelease={helmrelease} />