+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Upgrade backstage to v1.23.0
+- Upgrade backstage to v1.23.3
 - GS plugin: Use Kubernetes plugin to access clusters' APIs.
 - GS plugin: Add links to GitOps UI.
 - Flux plugin: Add links to GitOps UI.